requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
-e git://github.com/conda/conda.git#egg=conda
coverage==4.5.3
cryptography==2.7
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.19
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
libarchive-c==2.8
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.0
PySocks==1.7.0
pytest==5.0.1
pytest-cov==2.7.1
pytest-flask==0.15.0
pytest-mock==1.10.4
PyYAML==5.1.1
requests==2.22.0
ruamel-yaml==0.15.46
six==1.12.0
toml==0.10.0
toolz==0.10.0
tqdm==4.32.1
urllib3==1.24.2
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.2
